support SwiftPM
