+++ Platforms/iOS/GLHKView.swift
@@ -1,3 +1,5 @@
+#if os(iOS)
+
 import AVFoundation
 import GLKit
 
@@ -77,3 +79,5 @@ extension GLHKView: NetStreamDrawable {
         }
     }
 }
+
+#endif
