+++ Sources/Extension/CVPixelBuffer+Extension.swift
@@ -1,4 +1,5 @@
 import Foundation
+import CoreVideo
 
 extension CVPixelBuffer {
     var width: Int {
