+++ Sources/FLV/FLVTagType.swift
@@ -1,3 +1,5 @@
+import Foundation
+
 public enum FLVTagType: UInt8 {
     case audio = 8
     case video = 9
