+++ Sources/HTTP/M3U.swift
@@ -1,3 +1,5 @@
+import Foundation
+
 /**
  - seealso: https://tools.ietf.org/html/draft-pantos-http-live-streaming-19
  */
