+++ Sources/SwiftPMSupport/dummy.m
@@ -1,0 +1,3 @@
+#import "HaishinKit.h"
+
+void HaishinKit_SwiftPMSupport_dummy_symbol() {}
