+++ Sources/Util/Constants.swift
@@ -1,3 +1,7 @@
 import Logboard
 
+#if canImport(SwiftPMSupport)
+import SwiftPMSupport
+#endif
+
 let logger = Logboard.with(HaishinKitIdentifier)
