+++ Sources/Util/DataConvertible.swift
@@ -1,3 +1,5 @@
+import Foundation
+
 protocol DataConvertible {
     var data: Data { get set }
 }
